+++ core/providers/utils/utils.go
@@ -738,6 +738,11 @@ func GetProviderName(defaultProvider schemas.ModelProvider, customConfig *schema
 	return defaultProvider
 }
 
+// IsVertexAnthropicModel checks if the model is an Anthropic model in Vertex.
+func IsVertexAnthropicModel(model string) bool {
+	return strings.Contains(model, "claude")
+}
+
 // IsVertexMistralModel checks if the model is a Mistral or Codestral model in Vertex.
 func IsVertexMistralModel(model string) bool {
 	return strings.Contains(model, "mistral") || strings.Contains(model, "codestral")